controller/requirements.txt
redis
redis[hiredis]
pymongo
pymongo[srv]
python-dotenv